+++ src/Products/PloneMeeting/browser/overrides.py
@@ -16,7 +16,11 @@ from collective.eeafaceted.dashboard.browser.overrides import DashboardDocumentG
 from collective.eeafaceted.dashboard.browser.overrides import DashboardDocumentGeneratorLinksViewlet
 from collective.eeafaceted.dashboard.browser.views import RenderTermPortletView
 from collective.iconifiedcategory import utils as collective_iconifiedcategory_utils
+from collective.iconifiedcategory.browser.actionview import BaseView as BaseActionView
 from collective.iconifiedcategory.browser.actionview import ConfidentialChangeView
+from collective.iconifiedcategory.browser.actionview import PublishableChangeView
+from collective.iconifiedcategory.browser.actionview import SignedChangeView
+from collective.iconifiedcategory.browser.actionview import ToPrintChangeView
 from collective.iconifiedcategory.browser.tabview import CategorizedTabView
 from collective.iconifiedcategory.browser.views import CategorizedChildInfosView
 from collective.iconifiedcategory.browser.views import CategorizedChildView
@@ -53,6 +57,7 @@ from Products.CMFPlone.utils import safe_unicode
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from Products.PloneMeeting import utils as pm_utils
 from Products.PloneMeeting.config import BARCODE_INSERTED_ATTR_ID
+from Products.PloneMeeting.config import HAS_CAT_PUBLISHABLE
 from Products.PloneMeeting.config import ITEM_SCAN_ID_NAME
 from Products.PloneMeeting.interfaces import IMeeting
 from Products.PloneMeeting.utils import get_annexes
@@ -60,13 +65,6 @@ from Products.PloneMeeting.utils import sendMail
 from zope.annotation import IAnnotations
 from zope.i18n import translate
 from zope.interface import alsoProvides
-
-
-try:
-    HAS_CAT_PUBLISHABLE = True
-    from collective.iconifiedcategory.interfaces import ICategorizedPublishable
-except ImportError:
-    HAS_CAT_PUBLISHABLE = False
 
 
 class PMFolderContentsView(FolderContentsView):
@@ -1071,27 +1069,47 @@ class CategorizedAnnexesView(CategorizedTabView):
         super(CategorizedAnnexesView, self).__init__(context, request)
         self.portal_url = api.portal.get().absolute_url()
         self.tool = api.portal.get_tool('portal_plonemeeting')
+        self.cfg = self.tool.getMeetingConfig(self.context)
 
-    def _prepare_table_render(self, table, portal_type):
-        if portal_type == 'annexDecision':
-            self.request.set('force_use_item_decision_annexes_group', True)
-            self.config = collective_iconifiedcategory_utils.get_config_root(self.context)
-            self.request.set('force_use_item_decision_annexes_group', False)
-        else:
-            self.config = collective_iconifiedcategory_utils.get_config_root(self.context)
+    def table_render(self, portal_type=None):
+        """To be removed when using collective.iconifiedcategory 0.39+"""
+        self.portal_type = portal_type
+        return super(CategorizedAnnexesView, self).table_render(portal_type=portal_type)
 
-        if self.config.to_be_printed_activated:
-            alsoProvides(table, ICategorizedPrint)
-        if self.config.confidentiality_activated and self._showConfidentialColumn():
+    def _prepare_table_render(self, table):
+        """To be removed when using collective.iconifiedcategory 0.39+"""
+        if self.show('confidentiality'):
             alsoProvides(table, ICategorizedConfidential)
-        if HAS_CAT_PUBLISHABLE and self.config.publishable_activated:
+        if self.show('to_be_printed'):
+            alsoProvides(table, ICategorizedPrint)
+        if HAS_CAT_PUBLISHABLE and self.show('publishable'):
+            from collective.iconifiedcategory.interfaces import ICategorizedPublishable
             alsoProvides(table, ICategorizedPublishable)
-        if self.config.signed_activated:
+        if self.show('signed'):
             alsoProvides(table, ICategorizedSigned)
 
-    def _showConfidentialColumn(self):
+    def _config(self):
         """ """
-        return self.tool.isManager(self.context)
+        if self.portal_type == 'annexDecision':
+            self.request.set('force_use_item_decision_annexes_group', True)
+            config = collective_iconifiedcategory_utils.get_config_root(self.context)
+            self.request.set('force_use_item_decision_annexes_group', False)
+        else:
+            config = collective_iconifiedcategory_utils.get_config_root(self.context)
+        return config
+
+    def show(self, action_type):
+        """To be removed when using collective.iconifiedcategory 0.39+"""
+        config = self._config()
+        attr_config = '{0}_activated'.format(action_type)
+        show = getattr(config, attr_config) and self._show_column(action_type)
+        return show
+
+    def _show_column(self, action_type):
+        """Made to be overrided."""
+        annex_attr_config = '{0}_display'.format(action_type)
+        check = annex_attr_config in self.cfg.getAnnexRestrictShownAndEditableAttributes()
+        return not check or self.tool.isManager(self.context)
 
     def showAddAnnex(self):
         """ """
@@ -1177,15 +1195,11 @@ class PMCategorizedChildInfosView(CategorizedChildInfosView):
         """ """
         super(PMCategorizedChildInfosView, self).__init__(context, request)
         self.tool = api.portal.get_tool('portal_plonemeeting')
+        self.cfg = self.tool.getMeetingConfig(self.context)
 
     def show_preview_link(self):
         """Show link if preview is enabled, aka the auto_convert in collective.documentviewer."""
         return self.tool.auto_convert_annexes()
-
-    def show_confidential(self, element):
-        """ """
-        show = super(PMCategorizedChildInfosView, self).show_confidential(element)
-        return show and self.tool.isManager(self.context)
 
     def show_nothing(self):
         """Do not display the 'Nothing' label."""
@@ -1195,17 +1209,53 @@ class PMCategorizedChildInfosView(CategorizedChildInfosView):
         """ """
         return "{0}/@@categorized-annexes".format(self.context.absolute_url())
 
+    def show(self, element, attr_prefix):
+        """To be removed when using collective.iconifiedcategory 0.39+"""
+        show = element['{0}_activated'.format(attr_prefix)] and self._show_detail(attr_prefix)
+        if show:
+            self.have_details_to_show = True
+        return show
 
-class PMConfidentialChangeView(ConfidentialChangeView):
-    """Only available to Managers."""
+    def _show_detail(self, detail_type):
+        """ """
+        annex_attr_config = '{0}_display'.format(detail_type)
+        check = annex_attr_config in self.cfg.getAnnexRestrictShownAndEditableAttributes()
+        return not check or self.tool.isManager(self.context)
 
-    def _may_set_values(self, values):
-        res = super(PMConfidentialChangeView, self)._may_set_values(values)
+
+class PMBaseActionView(BaseActionView):
+    """ """
+    def _may_set_values(self, values, ):
+        res = super(PMBaseActionView, self)._may_set_values(values)
         if res:
-            # user must be MeetingManager
+            # get current type of action
+            # turn to_be_printed_activated to to_be_printed_edit/to_be_printed_display
+            # if an element is not displayed only to MeetingManagers, we consider it is ony editable as well
+            config_attr_display = self.category_group_attr_name.replace('_activated', '_display')
+            config_attr_edit = self.category_group_attr_name.replace('_activated', '_edit')
+            # restricted to MeetingManagers?
             tool = api.portal.get_tool('portal_plonemeeting')
-            res = tool.isManager(self.context)
+            cfg = tool.getMeetingConfig(self.context)
+            annex_attr_config = cfg.getAnnexRestrictShownAndEditableAttributes()
+            if config_attr_display in annex_attr_config or config_attr_edit in annex_attr_config:
+                res = tool.isManager(self.context)
         return res
+
+
+class PMConfidentialChangeView(ConfidentialChangeView, PMBaseActionView):
+    """Override to use new base klass."""
+
+
+class PMPublishableChangeView(PublishableChangeView, PMBaseActionView):
+    """Override to use new base klass."""
+
+
+class PMSignedChangeView(SignedChangeView, PMBaseActionView):
+    """Override to use new base klass."""
+
+
+class PMToPrintChangeView(ToPrintChangeView, PMBaseActionView):
+    """Override to use new base klass."""
 
 
 class PMReferenceBrowserPopup(ReferenceBrowserPopup):